convert server to one of tasks
