+++ app/cmd/main.go
@@ -4,16 +4,12 @@ import (
 	"context"
 	"flag"
 	"log"
-	"net/http"
 
-	"github.com/friendsofgo/errors"
 	"github.com/ktakenaka/gomsx/app/cmd/db"
+	"github.com/ktakenaka/gomsx/app/cmd/httpserver"
 	"github.com/ktakenaka/gomsx/app/cmd/logger"
 	"github.com/ktakenaka/gomsx/app/cmd/shutdown"
 	"github.com/ktakenaka/gomsx/app/config"
-
-	"github.com/labstack/echo/v4"
-	"github.com/labstack/echo/v4/middleware"
 )
 
 func main() {
@@ -42,24 +38,11 @@ func main() {
 	}
 	shutdownTaks.Add(dbTask)
 
-	// Server
-	e := echo.New()
-	e.Use(middleware.Logger())
-	e.Use(middleware.Recover())
-	e.GET("/", hello)
-	go func() {
-		if err := e.Start(":8080"); err != nil && err != http.ErrServerClosed {
-			e.Logger.Fatal(errors.Wrap(err, "Server failure"))
-		}
-	}()
+	serverTask, err := httpserver.Initialize(ctx, conf.App)
+	if err != nil {
+		log.Fatal(err)
+	}
+	shutdownTaks.Add(serverTask)
 
 	shutdownTaks.WaitForServerStop(ctx)
-	if err := e.Shutdown(ctx); err != nil {
-		e.Logger.Fatal(err)
-	}
-}
-
-// Handler
-func hello(c echo.Context) error {
-	return c.String(http.StatusOK, "Hello, World!")
 }
