+++ app/pkg/sqls/sqls.go
@@ -1,9 +1,9 @@
 package sqls
 
 import (
+	"fmt"
 	"time"
 
-	"github.com/friendsofgo/errors"
 	_ "github.com/go-sql-driver/mysql"
 	"github.com/jmoiron/sqlx"
 )
@@ -22,7 +22,7 @@ func Connect(conf *Config) (*DB, error) {
 	db, err := sqlx.Connect(conf.Driver, conf.ToConnString())
 	// TODO: retry severail times
 	if err != nil {
-		return nil, errors.Wrap(err, "failed to connect to DB")
+		return nil, fmt.Errorf("failed to connect to DB: %w", err)
 	}
 
 	if conf.MaxIdleConns == 0 {
